ref(login): Move main screen to own file
